+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [4.5.0] - 2026-05-16
+
+* Remove FontAwesome dependency.
+* Remove IconHelper class.
+* Fix DecimalValidator parse bug.
+* Fix DecimalTextFormatter format bug.
+
 ## [4.4.1] - 2026-02-22
 
 * Add method sort in ListField.